# Proposal for faster iteration over the weights of edges