@sbromberger 2019-02-10
This is fantastic work. SWG was written in a hurry and there are probably lots of opportunities for improved efficiency. PRs are very welcome.