@simonschoelly 2019-11-24
So, this was more a proof of concept.I think, it should be possible, to implement a more general solution in LightGraphs,which would lead to better result not only for `SimpleWeightedGraph` but also for `SimpleGraph` if a sparse matrix is used for weights. (Also, my own [package](https://github.com/simonschoelly/SimpleValueGraphs.jl) would work with this, if I ever find the time to finish it.).

So I think we can close this for now.